+++ R/extract_estimates.R
@@ -79,6 +79,7 @@ extract_estimates <- function(
     } 
   }
 
+
   # tidy_custom
   est_custom <- tidy_custom(model)
   sanity_tidy(est, est_custom, estimate, statistic, class(model)[1])
@@ -137,10 +138,23 @@ extract_estimates <- function(
   }
 
 
-  # round everything
+  # are statistics available? if not, display an informative error message
+  # check this after all custom statistics and stars are added
+  estimate_glue_strip <- regmatches(estimate_glue, gregexpr("\\{[^\\}]*\\}", estimate_glue))
+  estimate_glue_strip <- sort(unique(unlist(estimate_glue_strip)))
+  estimate_glue_strip <- gsub("\\{|\\}", "", estimate_glue_strip)
+  estimate_glue_strip <- setdiff(estimate_glue_strip, colnames(est))
+  if (length(estimate_glue_strip) > 0) {
+    stop(sprintf("These estimates or statistics do not seem to be available: %s. You can use the `get_estimates` function to see which statistics are available.", 
+                 paste(estimate_glue_strip, collapse = ", ")))
+  }
+
+
+  # round everything: ensures that the reshape doesn't produce incompatible types
   for (n in colnames(est)) {
     est[[n]] <- rounding(est[[n]], fmt)
   }
+
 
   # extract estimates (there can be several)
   for (i in seq_along(estimate_glue)) {